Moved command stuff into a separate file
