+++ src/CommandHandler.cpp
@@ -1,0 +1,71 @@
+#include "CommandHandler.h"
+#include <nlohmann/json.hpp>
+#include "helper.h"
+#include "json.h"
+#include "logger.h"
+#include "configuration.h"
+
+namespace BrightnESS {
+namespace ForwardEpicsToKafka {
+
+ConfigCB::ConfigCB(Main &main) : main(main) {}
+
+void ConfigCB::operator()(std::string const &msg) {
+  LOG(7, "Command received: {}", msg);
+  try {
+    handleCommand(msg);
+  } catch (...) {
+    LOG(3, "Command does not look like valid json: {}", msg);
+  }
+}
+
+void ConfigCB::handleCommandAdd(nlohmann::json const &Document) {
+  // Use instance of Configuration to extract stream info.
+  Configuration Config;
+  Config.extractStreamSettings();
+
+  for (auto & Stream : Config.StreamsInfo) {
+    main.addMapping(Stream);
+  }
+}
+
+void ConfigCB::handleCommandStopChannel(nlohmann::json const &Document) {
+  if (auto ChannelMaybe = find<std::string>("channel", Document)) {
+    main.streams.channel_stop(ChannelMaybe.inner());
+  }
+}
+
+void ConfigCB::handleCommandStopAll() { main.streams.streams_clear(); }
+
+void ConfigCB::handleCommandExit() { main.stopForwarding(); }
+
+void ConfigCB::handleCommand(std::string const &Msg) {
+  using nlohmann::json;
+  auto Document = json::parse(Msg);
+
+  std::string Command = findCommand(Document);
+
+  if (Command == "add") {
+    handleCommandAdd(Document);
+  } else if (Command == "stop_channel") {
+    handleCommandStopChannel(Document);
+  } else if (Command == "stop_all") {
+    handleCommandStopAll();
+  } else if (Command == "exit") {
+    handleCommandExit();
+  } else {
+    LOG(6, "Cannot understand command: {}", Command);
+  }
+}
+
+std::string ConfigCB::findCommand(nlohmann::json const &Document) {
+  if (auto CommandMaybe = find<std::string>("cmd", Document)) {
+    return CommandMaybe.inner();
+  }
+
+  return std::string();
+}
+
+}
+}
+
